Factor out Autosuggestions into a component
