+++ components/registration/index.js
@@ -1,4 +1,5 @@
 import { PatientRegistrationForm } from './PatientRegistrationForm';
 import { PatientInfo } from './PatientInfo';
+import { RegistrationAutoSuggest } from './RegistrationAutoSuggest';
 
-export { PatientRegistrationForm, PatientInfo };
+export { PatientRegistrationForm, PatientInfo, RegistrationAutoSuggest };
