+++ pages/registration/index.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import Autosuggest from 'react-autosuggest';
+// import Autosuggest from 'react-autosuggest';
 import toast from 'react-hot-toast';
 import moment from 'moment';
 import axiosInstance from '@/pages/api/_axiosInstance';
@@ -10,6 +10,7 @@ import withAuth from '@/utils/auth';
 import {
   PatientRegistrationForm,
   PatientInfo,
+  RegistrationAutoSuggest,
 } from '@/components/registration';
 import { Button } from '@/components/TextComponents/';
 import { useLoading } from '@/context/LoadingContext';
@@ -154,82 +155,82 @@ const Registration = () => {
   };
 
   // Auto Suggestions functions
-  const renderSuggestion = suggestion => {
-    const name = suggestion.name;
-    const id = `${suggestion.village_prefix} ${suggestion.pk
-      .toString()
-      .padStart(3, '0')}`;
-    const imageURL = suggestion.picture;
+  // const renderSuggestion = suggestion => {
+  //   const name = suggestion.name;
+  //   const id = `${suggestion.village_prefix} ${suggestion.pk
+  //     .toString()
+  //     .padStart(3, '0')}`;
+  //   const imageURL = suggestion.picture;
 
-    return (
-      <div
-        className="card cursor-pointer grid grid-cols-2"
-        onClick={() => setPatient(suggestion)}
-      >
-        <div className="self-center">
-          <img
-            src={`${CLOUDINARY_URL}/${imageURL}`}
-            alt="Placeholder image"
-            className="object-cover h-28 w-28 my-2"
-          />
-        </div>
+  //   return (
+  //     <div
+  //       className="card cursor-pointer grid grid-cols-2"
+  //       onClick={() => setPatient(suggestion)}
+  //     >
+  //       <div className="self-center">
+  //         <img
+  //           src={`${CLOUDINARY_URL}/${imageURL}`}
+  //           alt="Placeholder image"
+  //           className="object-cover h-28 w-28 my-2"
+  //         />
+  //       </div>
 
-        <div className="flex flex-col justify-center ml-2">
-          <div className="text-s font-medium text-gray-900">{id}</div>
-          <div className="text-s font-medium text-gray-900">{name}</div>
-        </div>
-      </div>
-    );
-  };
+  //       <div className="flex flex-col justify-center ml-2">
+  //         <div className="text-s font-medium text-gray-900">{id}</div>
+  //         <div className="text-s font-medium text-gray-900">{name}</div>
+  //       </div>
+  //     </div>
+  //   );
+  // };
 
-  // When suggestion is clicked, Autosuggest needs to populate the input
-  // based on the clicked suggestion. Teach Autosuggest how to calculate the
-  // input value for every given suggestion.
-  const getSuggestionValue = suggestion => {
-    return suggestion.name;
-  };
+  // // When suggestion is clicked, Autosuggest needs to populate the input
+  // // based on the clicked suggestion. Teach Autosuggest how to calculate the
+  // // input value for every given suggestion.
+  // const getSuggestionValue = suggestion => {
+  //   return suggestion.name;
+  // };
 
-  const onChange = (event, { newValue }) => {
-    setValue(newValue);
-  };
+  // const onChange = (event, { newValue }) => {
+  //   setValue(newValue);
+  // };
 
-  // Autosuggest will call this function every time you need to update suggestions.
-  const onSuggestionsFetchRequested = ({ value }) => {
-    const inputValue = value.trim().toLowerCase();
-    const query =
-      inputValue.length === 0
-        ? []
-        : patientsList.filter(patient =>
-            patient.filter_string.toLowerCase().includes(inputValue)
-          );
+  // // Autosuggest will call this function every time you need to update suggestions.
+  // const onSuggestionsFetchRequested = ({ value }) => {
+  //   const inputValue = value.trim().toLowerCase();
+  //   const query =
+  //     inputValue.length === 0
+  //       ? []
+  //       : patientsList.filter(patient =>
+  //           patient.filter_string.toLowerCase().includes(inputValue)
+  //         );
 
-    setSuggestions(query);
-  };
+  //   setSuggestions(query);
+  // };
 
-  // Autosuggest will call this function every time you need to clear suggestions.
-  const onSuggestionsClearRequested = () => {
-    setSuggestions([]);
-  };
+  // // Autosuggest will call this function every time you need to clear suggestions.
+  // const onSuggestionsClearRequested = () => {
+  //   setSuggestions([]);
+  // };
 
-  // Autosuggest will pass through all these props to the input.
-  const inputProps = {
-    placeholder: 'Search Patient',
-    type: 'search',
-    value,
-    onChange: onChange,
-    className:
-      'block w-full rounded-md border-0 py-1.5 px-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6',
-  };
+  // // Autosuggest will pass through all these props to the input.
+  // const inputProps = {
+  //   placeholder: 'Search Patient',
+  //   type: 'search',
+  //   value,
+  //   onChange: onChange,
+  //   className:
+  //     'block w-full rounded-md border-0 py-1.5 px-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6',
+  // };
 
-  const autosuggestTheme = {
-    container: 'react-autosuggest__container w-full',
-    input: 'react-autosuggest__input form-input w-full',
-    suggestionsContainer:
-      'react-autosuggest__suggestions-container w-full mt-2 border border-gray-300 rounded-md',
-    suggestionsList: 'react-autosuggest__suggestions-list w-full space-y-2 p-2',
-    suggestion:
-      'react-autosuggest__suggestion block w-full hover:bg-blue-100 transition-colors duration-300 ease-in-out p-2 border border-gray-300 rounded-md',
-  };
+  // const autosuggestTheme = {
+  //   container: 'react-autosuggest__container w-full',
+  //   input: 'react-autosuggest__input form-input w-full',
+  //   suggestionsContainer:
+  //     'react-autosuggest__suggestions-container w-full mt-2 border border-gray-300 rounded-md',
+  //   suggestionsList: 'react-autosuggest__suggestions-list w-full space-y-2 p-2',
+  //   suggestion:
+  //     'react-autosuggest__suggestion block w-full hover:bg-blue-100 transition-colors duration-300 ease-in-out p-2 border border-gray-300 rounded-md',
+  // };
 
   return (
     <div className="mx-4">
@@ -253,7 +254,7 @@ const Registration = () => {
             Registration
           </h1>
           <div className="flex items-center justify-center mb-2 w-full">
-            <Autosuggest
+            {/* <Autosuggest
               suggestions={suggestions}
               onSuggestionsFetchRequested={onSuggestionsFetchRequested}
               onSuggestionsClearRequested={onSuggestionsClearRequested}
@@ -261,6 +262,10 @@ const Registration = () => {
               renderSuggestion={renderSuggestion}
               inputProps={inputProps}
               theme={autosuggestTheme}
+            /> */}
+            <RegistrationAutoSuggest
+              patientsList={patientsList}
+              setPatient={setPatient}
             />
           </div>
           <div className="flex items-center justify-center mb-6 gap-3">
